Fix some deprecations
