+++ shared/homeless/composer.json
@@ -18,7 +18,7 @@
         "mnvx/lowrapper": "^1.0",
         "phpoffice/phpspreadsheet": "^1.23",
         "sonata-project/datagrid-bundle": "^3.0",
-        "sonata-project/doctrine-extensions": "^1.0",
+        "sonata-project/doctrine-extensions": "^2.0",
         "sonata-project/doctrine-orm-admin-bundle": "^4.0",
         "sonata-project/exporter": "^2.0",
         "sonata-project/formatter-bundle": "^5.0",
